# apps/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenCharts API",
    "version": "0.1.0",
    "description": "Open-source charting and market analysis platform API"
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Local development" }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "getHealth",
        "tags": ["system"],
        "responses": {
          "200": {
            "description": "Server health status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "uptime": { "type": "number" },
                    "ws": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/symbols": {
      "get": {
        "summary": "List all symbols",
        "operationId": "getSymbols",
        "tags": ["symbols"],
        "responses": {
          "200": {
            "description": "List of symbols",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Symbol" } }
              }
            }
          }
        }
      }
    },
    "/api/symbols/search": {
      "get": {
        "summary": "Search symbols",
        "operationId": "searchSymbols",
        "tags": ["symbols"],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Matching symbols",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Symbol" } }
              }
            }
          }
        }
      }
    },
    "/api/bars/{symbol}/{resolution}": {
      "get": {
        "summary": "Get historical bars",
        "operationId": "getBars",
        "tags": ["bars"],
        "parameters": [
          { "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "resolution", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "integer" } },
          { "name": "to", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 300 } }
        ],
        "responses": {
          "200": {
            "description": "OHLCV bars",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bars": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/OHLCV" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auth token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "operationId": "register",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "username", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auth token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges": {
      "get": {
        "summary": "List connected exchanges",
        "operationId": "getExchanges",
        "tags": ["exchanges"],
        "responses": {
          "200": {
            "description": "Exchange statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ExchangeStatus" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add an exchange",
        "operationId": "addExchange",
        "tags": ["exchanges"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "symbols"],
                "properties": {
                  "id": { "type": "string" },
                  "symbols": { "type": "array", "items": { "type": "string" } },
                  "pollIntervalMs": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Exchange added" }
        }
      }
    },
    "/api/exchanges/orderbook/{symbol}": {
      "get": {
        "summary": "Aggregated order book",
        "operationId": "getOrderBook",
        "tags": ["exchanges"],
        "parameters": [
          { "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "depth", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Aggregated order book",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AggregatedOrderBook" }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/ticker/{symbol}": {
      "get": {
        "summary": "Aggregated VWAP ticker",
        "operationId": "getTicker",
        "tags": ["exchanges"],
        "parameters": [
          { "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "VWAP ticker",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AggregatedTicker" }
              }
            }
          }
        }
      }
    },
    "/api/calendar": {
      "get": {
        "summary": "Economic calendar events",
        "operationId": "getCalendarEvents",
        "tags": ["calendar"],
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "country", "in": "query", "schema": { "type": "string" } },
          { "name": "impact", "in": "query", "schema": { "type": "string", "enum": ["low", "medium", "high"] } }
        ],
        "responses": {
          "200": {
            "description": "Calendar events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/EconEvent" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshots": {
      "post": {
        "summary": "Create chart snapshot",
        "operationId": "createSnapshot",
        "tags": ["snapshots"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["symbol", "resolution", "chartType"],
                "properties": {
                  "symbol": { "type": "string" },
                  "resolution": { "type": "string" },
                  "chartType": { "type": "string" },
                  "indicators": { "type": "array" },
                  "title": { "type": "string" },
                  "imageData": { "type": "string" },
                  "chartState": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Snapshot created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "shareCode": { "type": "string" },
                    "url": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshots/{code}": {
      "get": {
        "summary": "Get snapshot by share code",
        "operationId": "getSnapshot",
        "tags": ["snapshots"],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Snapshot data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChartSnapshot" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/plugins": {
      "get": {
        "summary": "List plugins",
        "operationId": "getPlugins",
        "tags": ["plugins"],
        "responses": {
          "200": {
            "description": "Plugin list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plugins": { "type": "array", "items": { "$ref": "#/components/schemas/PluginInfo" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Register a plugin",
        "operationId": "registerPlugin",
        "tags": ["plugins"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PluginRegistration" }
            }
          }
        },
        "responses": {
          "201": { "description": "Plugin registered" }
        }
      }
    },
    "/api/transpile": {
      "post": {
        "summary": "Transpile PineScript",
        "operationId": "transpile",
        "tags": ["transpile"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["source", "target"],
                "properties": {
                  "source": { "type": "string" },
                  "target": { "type": "string", "enum": ["typescript", "python"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transpiled code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string" },
                    "version": { "type": "integer", "nullable": true },
                    "target": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Symbol": {
        "type": "object",
        "properties": {
          "ticker": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "exchange": { "type": "string" },
          "assetClass": { "type": "string" }
        }
      },
      "OHLCV": {
        "type": "object",
        "properties": {
          "time": { "type": "integer" },
          "open": { "type": "number" },
          "high": { "type": "number" },
          "low": { "type": "number" },
          "close": { "type": "number" },
          "volume": { "type": "number" }
        }
      },
      "ExchangeStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "connected": { "type": "boolean" },
          "symbols": { "type": "array", "items": { "type": "string" } },
          "lastUpdate": { "type": "integer", "nullable": true },
          "latencyMs": { "type": "integer", "nullable": true }
        }
      },
      "AggregatedOrderBook": {
        "type": "object",
        "properties": {
          "symbol": { "type": "string" },
          "timestamp": { "type": "integer" },
          "bids": { "type": "array", "items": { "$ref": "#/components/schemas/OrderBookLevel" } },
          "asks": { "type": "array", "items": { "$ref": "#/components/schemas/OrderBookLevel" } },
          "bestBid": { "$ref": "#/components/schemas/OrderBookLevel" },
          "bestAsk": { "$ref": "#/components/schemas/OrderBookLevel" },
          "vwap": { "type": "number", "nullable": true },
          "spread": { "type": "number", "nullable": true },
          "exchanges": { "type": "array", "items": { "type": "string" } }
        }
      },
      "OrderBookLevel": {
        "type": "object",
        "properties": {
          "price": { "type": "number" },
          "amount": { "type": "number" },
          "exchange": { "type": "string" }
        }
      },
      "AggregatedTicker": {
        "type": "object",
        "properties": {
          "symbol": { "type": "string" },
          "price": { "type": "number" },
          "prices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "exchange": { "type": "string" },
                "price": { "type": "number" },
                "volume": { "type": "number" }
              }
            }
          },
          "totalVolume": { "type": "number" },
          "timestamp": { "type": "integer" }
        }
      },
      "EconEvent": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "country": { "type": "string" },
          "date": { "type": "string", "format": "date-time" },
          "impact": { "type": "string", "enum": ["low", "medium", "high"] },
          "forecast": { "type": "string", "nullable": true },
          "previous": { "type": "string", "nullable": true },
          "actual": { "type": "string", "nullable": true },
          "currency": { "type": "string" }
        }
      },
      "ChartSnapshot": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "shareCode": { "type": "string" },
          "symbol": { "type": "string" },
          "resolution": { "type": "string" },
          "chartType": { "type": "string" },
          "indicators": { "type": "array" },
          "title": { "type": "string", "nullable": true },
          "imageData": { "type": "string", "nullable": true },
          "chartState": { "type": "object", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PluginInfo": {
        "type": "object",
        "properties": {
          "manifest": { "$ref": "#/components/schemas/PluginManifest" },
          "status": { "type": "string", "enum": ["active", "disabled", "error"] },
          "error": { "type": "string" }
        }
      },
      "PluginManifest": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "description": { "type": "string" },
          "author": { "type": "string" },
          "provides": { "type": "array", "items": { "type": "string" } }
        }
      },
      "PluginRegistration": {
        "type": "object",
        "required": ["manifest"],
        "properties": {
          "manifest": { "$ref": "#/components/schemas/PluginManifest" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}